@@ -6,7 +6,7 @@
 
     <groupId>io.jpower.kcp</groupId>
     <artifactId>kcp-parent</artifactId>
-    <version>1.5.0</version>
+    <version>1.4.13</version>
     <packaging>pom</packaging>
 
     <name>kcp-netty [Parent]</name>
@@ -81,17 +81,27 @@
                 <version>${slf4j.version}</version>
             </dependency>
 
+            <dependency>
+                <groupId>org.projectlombok</groupId>
+                <artifactId>lombok</artifactId>
+                <version>1.18.22</version>
+                <scope>provided</scope>
+            </dependency>
+
             <!-- test -->
             <dependency>
-                <groupId>junit</groupId>
-                <artifactId>junit</artifactId>
-                <version>4.13.1</version>
+                <groupId>org.junit.jupiter</groupId>
+                <artifactId>junit-jupiter-api</artifactId>
+                <version>5.8.1</version>
+                <scope>test</scope>
             </dependency>
+
             <dependency>
                 <groupId>org.easymock</groupId>
                 <artifactId>easymock</artifactId>
                 <version>3.5</version>
             </dependency>
+
         </dependencies>
     </dependencyManagement>
 
@@ -131,8 +141,8 @@
                 <groupId>org.apache.maven.plugins</groupId>
                 <artifactId>maven-compiler-plugin</artifactId>
                 <configuration>
-                    <source>1.8</source>
-                    <target>1.8</target>
+                    <source>11</source>
+                    <target>11</target>
                     <fork>true</fork>
                 </configuration>
             </plugin>
