@@ -1,14 +1,13 @@
 package io.jpower.kcp.netty.internal;
 
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertFalse;
-import static org.junit.Assert.assertTrue;
-import static org.junit.Assert.fail;
+
+import org.junit.jupiter.api.Test;
 
 import java.util.Collections;
 import java.util.NoSuchElementException;
 
-import org.junit.Test;
+import static org.junit.jupiter.api.Assertions.*;
+
 
 /**
  * @author <a href="mailto:szhnet@gmail.com">szh</a>
