@@ -21,7 +21,7 @@ public final class EchoClient {
 
     static final int CONV = Integer.parseInt(System.getProperty("conv", "10"));
     static final String HOST = System.getProperty("host", "127.0.0.1");
-    static final int PORT = Integer.parseInt(System.getProperty("port", "8007"));
+    static final int PORT = Integer.parseInt(System.getProperty("port", "9007"));
     static final int SIZE = Integer.parseInt(System.getProperty("size", "256"));
 
     public static void main(String[] args) throws Exception {
